Revert "finish what hannibal started"
This reverts commit 9658bf7ca826f6f5b946346da2d9b897af1d5c58.
