+++ src/main/java/at/hannibal2/skyhanni/features/garden/fortuneguide/CaptureFarmingGear.kt
@@ -387,13 +387,6 @@ object CaptureFarmingGear {
         }
     }
 
-    fun clearGearCommand() {
-        val storage = GardenAPI.storage?.fortune ?: return
-        ChatUtils.chat("clearing farming items")
-        storage.farmingItems.clear()
-        storage.outdatedItems.clear()
-    }
-
     @SubscribeEvent
     fun onConfigUpdaterMigratorConfigFix(event: ConfigUpdaterMigrator.ConfigFixEvent) {
         event.move(48, "#profile.garden.fortune.carrotFortune", "#profile.garden.fortune.carrolyn.CARROT")
